+++ CHANGES.md
@@ -41,11 +41,16 @@ To be released.  Note that 1.6.0 was skipped due to a mistake in the versioning.
      -  Added `HttpMessageSignaturesSpecDeterminer` interface.
      -  Added `--first-knock` option to `fedify lookup` command.
 
- -  Added `WorkersKvStore` class.  [[#233], [#241], [#242]]
+ -  Fedify now supports [Cloudflare Workers].  [[#233]]
+
+     -  Added `Federation.processQueuedTask()` method.  [[#242]]
+     -  Added `Message` type.  [[#242]]
+     -  Added `WorkersKvStore` class.  [[#241], [#242]]
 
  -  The minimum supported version of Node.js is now 22.0.0.
 
 [RFC 9421]: https://www.rfc-editor.org/rfc/rfc9421
+[Cloudflare Workers]: https://workers.cloudflare.com/
 [#208]: https://github.com/fedify-dev/fedify/issues/208
 [#227]: https://github.com/fedify-dev/fedify/issues/227
 [#233]: https://github.com/fedify-dev/fedify/issues/233